split code for graphql
